BOAC-1882 Move advising notes into their own merged model
