+++ boac/api/search_controller.py
@@ -31,7 +31,8 @@ from boac.externals.data_loch import get_enrolled_primary_sections
 from boac.lib import util
 from boac.lib.berkeley import current_term_id
 from boac.lib.http import tolerant_jsonify
-from boac.merged.student import search_advising_notes, search_for_students
+from boac.merged.advising_note import search_advising_notes
+from boac.merged.student import search_for_students
 from boac.models.alert import Alert
 from flask import current_app as app, request
 from flask_login import current_user, login_required
